Windows port: point_to_point: fix MSVC narrowing conversions
- device/host/point_to_point_device_op.cpp: static_cast std::abs(int)/int
  hop counts to uint32_t in the aggregate return.
- device/host/receive_program_factory.cpp, send_program_factory.cpp:
  static_cast size_t/DeviceAddr values (packet header size, buffer and
  semaphore addresses) down to uint32_t at CBDescriptor/CBFormatDescriptor
  brace-inits and runtime-args vector literals.

+++ ttnn/cpp/ttnn/operations/point_to_point/device/host/point_to_point_device_op.cpp
@@ -90,11 +90,11 @@ Fabric1DRoute fabric_1d_routing(
             bool ring_is_forward = (ring_hops > 0);
 
             const auto next_fabric_id = get_neighbor_id(ring_is_forward, MeshCoordinate::BoundaryMode::WRAP);
-            return {std::abs(ring_hops), !ring_is_forward, next_fabric_id};
+            return {static_cast<uint32_t>(std::abs(ring_hops)), !ring_is_forward, next_fabric_id};
         }
     }
     const auto next_fabric_id = get_neighbor_id(line_is_forward, MeshCoordinate::BoundaryMode::NONE);
-    return {line_hops, !line_is_forward, next_fabric_id};
+    return {static_cast<uint32_t>(line_hops), !line_is_forward, next_fabric_id};
 }
 }  // namespace detail
